Generate time log report even when process ID is unknown
The time log report demonstrates an example of mapped process phases, by
retrieving the process ID of the current project and requesting that all
phases be mapped to that underlying process. Unfortunately, this
approach does not work if we are querying data from more than one
project, since there is no "global" process ID. In that scenario, fall
back and display a simple, unmapped phase name.